# RuGram/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  dto.WhoamiResponse:
    properties:
      email:
        type: string
      id:
        type: string
      phone:
        type: string
    type: object
  utils.ErrorResponseStruct:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:4200
info:
  contact:
    email: support@rugram.com
    name: Gleb, Pavel
  description: API for RuGram social platform with MongoDB, JWT, OAuth, Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RuGram API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns access/refresh tokens in cookies
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes current access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - Auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: Revokes all access and refresh tokens for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Logout from all sessions
      tags:
      - Auth
...
  /auth/oauth/vk/callback:
    get:
      description: Handles VK OAuth callback, creates/authenticates user
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      summary: VK OAuth callback
      tags:
      - Auth
...
  /auth/oauth/yandex/callback:
    get:
      description: Handles Yandex OAuth callback, creates/authenticates user
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      summary: Yandex OAuth callback
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Uses refresh token from cookie to issue new access/refresh tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      summary: Refresh tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      summary: Register a new user
      tags:
      - Auth
  /auth/whoami:
    get:
      consumes:
      - application/json
      description: Returns authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WhoamiResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"